fixed error with the object using wrong kin data... added a set position function to objects... added update functionality, rendering functionality, and creation functionality to the object manager... fixed a small "error" within the BK_Vector...

+++ animal3D-SDK/animal3D-SDK/source/animal3D-DemoPlugin/A3_DEMO/BK_Vector.h
@@ -24,7 +24,7 @@ struct BK_Vector2
 		yVal = newYVal;
 	}
 
-	float magnitude() { return sqrt(xVal * xVal + yVal * yVal); }
+	float magnitude() { return (float)sqrt( (double)xVal * (double)xVal + (double) yVal * (double)yVal); }
 
 	float Dot(BK_Vector2 other)
 	{
